Script to find (and delete if requested) TNMR backup files
There is an option in the TNMR control software to save backup files at
regular intervals during data acquisition, so that in the event of power
failure or system crash only the data acquired since the last backup file is
lost. These backup files are saved in the same directory as the main file and
have the same name but with a suffix of _N (where N is a decimal number) on
the basename.

It is possible to use a standard find command to find files ending in _N.tnt,
but this will also catch files which just happen to have been given a name
ending in _N, which are not backup files. This script adds some sanity checks
to avoid finding files which look like backup files but actually aren't.